DB functions, Cache implementation

+++ src/main/kotlin/dev/notmarra/kilevels/utils/config/MainConfig.kt
@@ -36,7 +36,7 @@ data class DataPool(
 
 @ConfigSerializable
 data class CacheConfig(
-    val ttl: Int = 300,
+    val ttl: Long = 300,
     val destroyOnLeave: Boolean = true,
 )
